helper: add scaleutils to provde cluster scale in helper tasks.
When performing a scale in action on a pool of Nomad nodes, there
are a number of tasks which should be conducted before the
instances are terminated within the remote infrastructure provider.

The scaleutils helper package provides a method for performing
Nomad node identification, node draining, node drain monitoring
and nodeID translation. This is exposed as a single function so
that target plugin authors can consume the functionality in a
simple manner.