policy: add new file source for loading policies from disk.
The MVP cluster scaling includes the ability to load scaling
policies from a specified directory. This functionality is off by
default but is turned on when the policy-dir flag or corresponding
config file parameter is set to a non-empty string.

The file source implements the source interface. It is also
responsible for generating UUIDs for the policies, a task which has
before been performed by Nomad. The source stores the ID mapped to
the file path md5sum. This allows the policy contents to change
without the need to update the policyID.

Funcationality is in place to allow reloading of the policies dir
by sending the agent a SIGHUP signal. This still requires plumbing
in from the agent and will be addressed in a follow up PR.